refactor: revise user API service

+++ UserService/internal/service/user/delete.go
@@ -1,7 +1,19 @@
 package user
 
-import "context"
+import (
+	"context"
+	"user-service/internal/errorz"
+)
 
-func (s *Service) DeleteUser(ctx context.Context, id string) error {
+func (s *Service) Delete(ctx context.Context, id string) error {
+	userID, ok := ctx.Value("user_id").(string)
+	if !ok {
+		return errorz.ErrUserIDNotSet
+	}
+
+	if userID != id {
+		return errorz.ErrUserAccessDenied
+	}
+
 	return s.userRepository.Delete(ctx, id)
 }
